Brain/OpenCV/requirements.txt
opencv-python 
numpy 
onnxruntime-gpu 
ultralytics
timm
onnxscript
matplotlib